Partial fix.
Builds, but get 'application failed to start 0xc000007b' error.

Need to test back in office.

+++ Lab10_Exercise01/Exercise01.cu
@@ -57,7 +57,7 @@ __global__ void simple_instance_kernel(float4 *instance_data, unsigned int width
 
 void executeSimulation()
 {
-	unsigned int num_bytes;
+	size_t num_bytes;
 	float4 *dptr;
 
 	// Exercise 1.1) Map CUDA graphics resource
